Break FASTERAsync out to separate files
