+++ cs/src/core/Async/DeleteAsync.cs
@@ -1,0 +1,108 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+using System.Diagnostics;
+using System.Runtime.CompilerServices;
+using System.Runtime.ExceptionServices;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace FASTER.core
+{
+    public partial class FasterKV<Key, Value> : FasterBase, IFasterKV<Key, Value>
+    {
+        /// <summary>
+        /// Contained state storage for the completion of an async Delete, or the result if the Delete was completed synchronously
+        /// </summary>
+        public struct DeleteAsyncResult<Input, Output, Context> : IUpdelAsyncResult<Input, Output, Context, DeleteAsyncResult<Input, Output, Context>>
+        {
+            private readonly OperationStatus internalStatus;
+            internal readonly UpdelAsyncInternal<Input, Output, Context, DeleteAsyncOperation<Input, Output, Context>, DeleteAsyncResult<Input, Output, Context>> updelAsyncInternal;
+
+            /// <summary>Current status of the Upsert operation</summary>
+            public Status Status => TranslateStatus(internalStatus);
+
+            internal DeleteAsyncResult(OperationStatus internalStatus)
+            {
+                Debug.Assert(internalStatus == OperationStatus.SUCCESS || internalStatus == OperationStatus.NOTFOUND);
+                this.internalStatus = internalStatus;
+                this.updelAsyncInternal = default;
+            }
+
+            internal DeleteAsyncResult(FasterKV<Key, Value> fasterKV, IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+                FasterExecutionContext<Input, Output, Context> currentCtx, PendingContext<Input, Output, Context> pendingContext, ExceptionDispatchInfo exceptionDispatchInfo)
+            {
+                internalStatus = OperationStatus.ALLOCATE_FAILED;
+                updelAsyncInternal = new UpdelAsyncInternal<Input, Output, Context, DeleteAsyncOperation<Input, Output, Context>, DeleteAsyncResult<Input, Output, Context>>(
+                                        fasterKV, fasterSession, currentCtx, pendingContext, exceptionDispatchInfo);
+            }
+
+            /// <summary>Complete the Delete operation, issuing additional allocation asynchronously if needed. It is usually preferable to use Complete() instead of this.</summary>
+            /// <returns>ValueTask for Delete result. User needs to await again if result status is Status.PENDING.</returns>
+            public ValueTask<DeleteAsyncResult<Input, Output, Context>> CompleteAsync(CancellationToken token = default)
+            {
+                if (internalStatus != OperationStatus.ALLOCATE_FAILED)
+                    return new ValueTask<DeleteAsyncResult<Input, Output, Context>>(new DeleteAsyncResult<Input, Output, Context>(internalStatus));
+                return updelAsyncInternal.CompleteAsync(token);
+            }
+
+            /// <summary>Complete the Delete operation, issuing additional I/O synchronously if needed.</summary>
+            /// <returns>Status of Delete operation</returns>
+            public Status Complete()
+            {
+                if (internalStatus != OperationStatus.ALLOCATE_FAILED)
+                    return this.Status;
+                return updelAsyncInternal.Complete();
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal ValueTask<DeleteAsyncResult<Input, Output, Context>> DeleteAsync<Input, Output, Context>(IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+            FasterExecutionContext<Input, Output, Context> currentCtx, ref Key key, Context userContext, long serialNo, CancellationToken token = default)
+        {
+            var pcontext = default(PendingContext<Input, Output, Context>);
+            pcontext.IsAsync = true;
+            return DeleteAsync(fasterSession, currentCtx, ref pcontext, ref key, userContext, serialNo, token);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal ValueTask<DeleteAsyncResult<Input, Output, Context>> DeleteAsync<Input, Output, Context>(IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+            FasterExecutionContext<Input, Output, Context> currentCtx, ref PendingContext<Input, Output, Context> pcontext, ref Key key, Context userContext, long serialNo, CancellationToken token)
+        {
+            CompletionEvent flushEvent;
+
+            fasterSession.UnsafeResumeThread();
+            try
+            {
+                OperationStatus internalStatus;
+                do
+                {
+                    flushEvent = hlog.FlushEvent;
+                    internalStatus = InternalDelete(ref key, ref userContext, ref pcontext, fasterSession, currentCtx, serialNo);
+                } while (internalStatus == OperationStatus.RETRY_NOW);
+
+                if (internalStatus == OperationStatus.SUCCESS || internalStatus == OperationStatus.NOTFOUND)
+                    return new ValueTask<DeleteAsyncResult<Input, Output, Context>>(new DeleteAsyncResult<Input, Output, Context>(internalStatus));
+                Debug.Assert(internalStatus == OperationStatus.ALLOCATE_FAILED);
+            }
+            finally
+            {
+                Debug.Assert(serialNo >= currentCtx.serialNum, "Operation serial numbers must be non-decreasing");
+                currentCtx.serialNum = serialNo;
+                fasterSession.UnsafeSuspendThread();
+            }
+
+            return SlowDeleteAsync(this, fasterSession, currentCtx, pcontext, flushEvent, token);
+        }
+
+        private static async ValueTask<DeleteAsyncResult<Input, Output, Context>> SlowDeleteAsync<Input, Output, Context>(
+            FasterKV<Key, Value> @this,
+            IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+            FasterExecutionContext<Input, Output, Context> currentCtx,
+            PendingContext<Input, Output, Context> pcontext, CompletionEvent flushEvent, CancellationToken token = default)
+        {
+            ExceptionDispatchInfo exceptionDispatchInfo = await WaitForFlushCompletionAsync(@this, currentCtx, flushEvent, token).ConfigureAwait(false);
+            return new DeleteAsyncResult<Input, Output, Context>(@this, fasterSession, currentCtx, pcontext, exceptionDispatchInfo);
+        }
+    }
+}
