+++ cs/src/core/Async/RMWAsync.cs
@@ -1,0 +1,282 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+using System;
+using System.Diagnostics;
+using System.Runtime.CompilerServices;
+using System.Runtime.ExceptionServices;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace FASTER.core
+{
+    public partial class FasterKV<Key, Value> : FasterBase, IFasterKV<Key, Value>
+    {
+        internal sealed class RmwAsyncInternal<Input, Output, Context>
+        {
+            const int Completed = 1;
+            const int Pending = 0;
+            ExceptionDispatchInfo _exception;
+            readonly FasterKV<Key, Value> _fasterKV;
+            readonly IFasterSession<Key, Value, Input, Output, Context> _fasterSession;
+            readonly FasterExecutionContext<Input, Output, Context> _currentCtx;
+            PendingContext<Input, Output, Context> _pendingContext;
+            AsyncIOContext<Key, Value> _diskRequest;
+            int CompletionComputeStatus;
+
+            internal RmwAsyncInternal(FasterKV<Key, Value> fasterKV, IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+                                      FasterExecutionContext<Input, Output, Context> currentCtx, PendingContext<Input, Output, Context> pendingContext,
+                                      AsyncIOContext<Key, Value> diskRequest, ExceptionDispatchInfo exceptionDispatchInfo)
+            {
+                _exception = exceptionDispatchInfo;
+                _fasterKV = fasterKV;
+                _fasterSession = fasterSession;
+                _currentCtx = currentCtx;
+                _pendingContext = pendingContext;
+                _diskRequest = diskRequest;
+                CompletionComputeStatus = Pending;
+            }
+
+            internal void SetDiskRequest(AsyncIOContext<Key, Value> diskRequest) => _diskRequest = diskRequest;
+
+            internal ValueTask<RmwAsyncResult<Input, Output, Context>> CompleteAsync(CancellationToken token = default)
+            {
+                Debug.Assert(_fasterKV.RelaxedCPR);
+
+                // Note: We currently do not await anything here, and we must never do any post-await work inside CompleteAsync; this includes any code in
+                // a 'finally' block. All post-await work must be re-initiated by end user on the mono-threaded session.
+
+                if (TryCompleteAsyncState(asyncOp: true, out AsyncIOContext<Key, Value> newDiskRequest, out CompletionEvent flushEvent, out var rmwAsyncResult))
+                    return new ValueTask<RmwAsyncResult<Input, Output, Context>>(rmwAsyncResult);
+
+                if (_exception != default)
+                    _exception.Throw();
+                return SlowRmwAsync(_fasterKV, _fasterSession, _currentCtx, _pendingContext, newDiskRequest, flushEvent, token);
+            }
+
+            internal bool TryCompleteAsyncState(bool asyncOp, out AsyncIOContext<Key, Value> newDiskRequest, out CompletionEvent flushEvent, out RmwAsyncResult<Input, Output, Context> rmwAsyncResult)
+            {
+                // This makes one attempt to complete the async operation's synchronous state, and clears the async pending counters.
+                if (CompletionComputeStatus != Completed
+                    && Interlocked.CompareExchange(ref CompletionComputeStatus, Completed, Pending) == Pending)
+                {
+                    try
+                    {
+                        if (_exception == default)
+                            return TryCompleteSync(asyncOp, out newDiskRequest, out flushEvent, out rmwAsyncResult);
+                    }
+                    catch (Exception e)
+                    {
+                        _exception = ExceptionDispatchInfo.Capture(e);
+                    }
+                    finally
+                    {
+                        _currentCtx.ioPendingRequests.Remove(_pendingContext.id);
+                        _currentCtx.asyncPendingCount--;
+                        _currentCtx.pendingReads.Remove();
+                    }
+                }
+
+                newDiskRequest = default;
+                flushEvent = default;
+                rmwAsyncResult = default;
+                return false;
+            }
+
+            internal bool TryCompleteSync(bool asyncOp, out AsyncIOContext<Key, Value> newDiskRequest, out CompletionEvent flushEvent, out RmwAsyncResult<Input, Output, Context> rmwAsyncResult)
+            {
+                _fasterSession.UnsafeResumeThread();
+                try
+                {
+                    Status status;
+                    if (!_diskRequest.IsDefault())
+                    {
+                        flushEvent = _fasterKV.hlog.FlushEvent;
+                        status = _fasterKV.InternalCompletePendingRequestFromContext(_currentCtx, _currentCtx, _fasterSession, _diskRequest, ref _pendingContext, asyncOp, out newDiskRequest);
+                    }
+                    else
+                    {
+                        status = _fasterKV.CallInternalRMW(_fasterSession, _currentCtx, ref _pendingContext, ref _pendingContext.key.Get(), ref _pendingContext.input.Get(), _pendingContext.userContext,
+                                                         _pendingContext.serialNum, asyncOp, out flushEvent, out newDiskRequest);
+                    }
+
+                    if (status != Status.PENDING)
+                    {
+                        _pendingContext.Dispose();
+                        rmwAsyncResult = new RmwAsyncResult<Input, Output, Context>(status, default);
+                        return true;
+                    }
+                }
+                finally
+                {
+                    _fasterSession.UnsafeSuspendThread();
+                }
+                rmwAsyncResult = default;
+                return false;
+            }
+
+            internal Status CompletePending()
+            {
+                _fasterSession.CompletePendingWithOutputs(out var completedOutputs, wait: true, spinWaitForCommit: false);
+                var status = completedOutputs.Next() ? completedOutputs.Current.Status : Status.ERROR;
+                completedOutputs.Dispose();
+                return status;
+            }
+        }
+
+        /// <summary>
+        /// State storage for the completion of an async RMW, or the result if the RMW was completed synchronously
+        /// </summary>
+        public struct RmwAsyncResult<Input, Output, Context>
+        {
+            /// <summary>Current status of the RMW operation</summary>
+            public Status Status { get; }
+
+            internal readonly Output output;
+
+            private readonly RmwAsyncInternal<Input, Output, Context> rmwAsyncInternal;
+
+            internal RmwAsyncResult(Status status, Output output)
+            {
+                this.Status = status;
+                this.output = output;
+                this.rmwAsyncInternal = default;
+            }
+
+            internal RmwAsyncResult(FasterKV<Key, Value> fasterKV, IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+                FasterExecutionContext<Input, Output, Context> currentCtx, PendingContext<Input, Output, Context> pendingContext,
+                AsyncIOContext<Key, Value> diskRequest, ExceptionDispatchInfo exceptionDispatchInfo)
+            {
+                Status = Status.PENDING;
+                output = default;
+                rmwAsyncInternal = new RmwAsyncInternal<Input, Output, Context>(fasterKV, fasterSession, currentCtx, pendingContext, diskRequest, exceptionDispatchInfo);
+            }
+
+            /// <summary>
+            /// Complete the RMW operation, issuing additional (rare) I/O asynchronously if needed.
+            /// It is usually preferable to use Complete() instead of this.
+            /// </summary>
+            /// <returns>ValueTask for RMW result. User needs to await again if result status is Status.PENDING.</returns>
+            public ValueTask<RmwAsyncResult<Input, Output, Context>> CompleteAsync(CancellationToken token = default)
+            {
+                if (Status != Status.PENDING)
+                    return new ValueTask<RmwAsyncResult<Input, Output, Context>>(new RmwAsyncResult<Input, Output, Context>(Status, default));
+                return rmwAsyncInternal.CompleteAsync(token);
+            }
+
+            /// <summary>
+            /// Complete the RMW operation, issuing additional (rare) I/O synchronously if needed.
+            /// </summary>
+            /// <returns>Status of RMW operation</returns>
+            public Status Complete()
+            {
+                if (Status != Status.PENDING)
+                    return Status;
+
+                if (!rmwAsyncInternal.TryCompleteAsyncState(asyncOp: false, out var newDiskRequest, out var flushEvent, out var rmwAsyncResult))
+                {
+                    if (newDiskRequest.IsDefault())
+                        flushEvent.Wait();
+                    rmwAsyncInternal.SetDiskRequest(newDiskRequest);
+
+                    while (!rmwAsyncInternal.TryCompleteSync(asyncOp: false, out newDiskRequest, out flushEvent, out rmwAsyncResult))
+                    {
+                        // CompletePending() may encounter OperationStatus.ALLOCATE_FAILED; if so, we don't have a more current flushEvent.
+                        if (!newDiskRequest.IsDefault() && rmwAsyncInternal.CompletePending() != Status.PENDING)
+                            break;
+                        flushEvent.Wait();
+                    }
+                }
+                return rmwAsyncResult.Status;
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal ValueTask<RmwAsyncResult<Input, Output, Context>> RmwAsync<Input, Output, Context>(IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+            FasterExecutionContext<Input, Output, Context> currentCtx, ref Key key, ref Input input, Context context, long serialNo, CancellationToken token = default)
+        {
+            var pcontext = default(PendingContext<Input, Output, Context>);
+            pcontext.IsAsync = true;
+            return RmwAsync(fasterSession, currentCtx, ref pcontext, ref key, ref input, context, serialNo, token);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal ValueTask<RmwAsyncResult<Input, Output, Context>> RmwAsync<Input, Output, Context>(IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+            FasterExecutionContext<Input, Output, Context> currentCtx, ref PendingContext<Input, Output, Context> pcontext, ref Key key, ref Input input, Context context, long serialNo, CancellationToken token)
+        {
+            var diskRequest = default(AsyncIOContext<Key, Value>);
+            CompletionEvent flushEvent;
+
+            fasterSession.UnsafeResumeThread();
+            try
+            {
+                var status = CallInternalRMW(fasterSession, currentCtx, ref pcontext, ref key, ref input, context, serialNo, asyncOp: true, out flushEvent, out diskRequest);
+                if (status != Status.PENDING)
+                    return new ValueTask<RmwAsyncResult<Input, Output, Context>>(new RmwAsyncResult<Input, Output, Context>(status, default));
+            }
+            finally
+            {
+                Debug.Assert(serialNo >= currentCtx.serialNum, "Operation serial numbers must be non-decreasing");
+                currentCtx.serialNum = serialNo;
+                fasterSession.UnsafeSuspendThread();
+            }
+
+            return SlowRmwAsync(this, fasterSession, currentCtx, pcontext, diskRequest, flushEvent, token);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private Status CallInternalRMW<Input, Output, Context>(IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+            FasterExecutionContext<Input, Output, Context> currentCtx, ref PendingContext<Input, Output, Context> pcontext, ref Key key, ref Input input, Context context, long serialNo,
+            bool asyncOp, out CompletionEvent flushEvent, out AsyncIOContext<Key, Value> diskRequest)
+        {
+            diskRequest = default;
+            OperationStatus internalStatus;
+            do
+            {
+                flushEvent = hlog.FlushEvent;
+                internalStatus = InternalRMW(ref key, ref input, ref context, ref pcontext, fasterSession, currentCtx, serialNo);
+            } while (internalStatus == OperationStatus.RETRY_NOW || internalStatus == OperationStatus.RETRY_LATER);
+
+            if (internalStatus == OperationStatus.SUCCESS || internalStatus == OperationStatus.NOTFOUND)
+                return (Status)internalStatus;
+            if (internalStatus == OperationStatus.ALLOCATE_FAILED)
+                return Status.PENDING;    // This plus diskRequest.IsDefault() means allocate failed
+
+            flushEvent = default;
+            return HandleOperationStatus(currentCtx, currentCtx, ref pcontext, fasterSession, internalStatus, asyncOp, out diskRequest);
+        }
+
+        private static async ValueTask<RmwAsyncResult<Input, Output, Context>> SlowRmwAsync<Input, Output, Context>(
+            FasterKV<Key, Value> @this, IFasterSession<Key, Value, Input, Output, Context> fasterSession,
+            FasterExecutionContext<Input, Output, Context> currentCtx, PendingContext<Input, Output, Context> pcontext,
+            AsyncIOContext<Key, Value> diskRequest, CompletionEvent flushEvent, CancellationToken token = default)
+        {
+            currentCtx.asyncPendingCount++;
+            currentCtx.pendingReads.Add();
+
+            ExceptionDispatchInfo exceptionDispatchInfo = default;
+            try
+            {
+                token.ThrowIfCancellationRequested();
+
+                if (@this.epoch.ThisInstanceProtected())
+                    throw new NotSupportedException("Async operations not supported over protected epoch");
+
+                // If we are here because of flushEvent, then _diskRequest is default--there is no pending disk operation.
+                if (diskRequest.IsDefault())
+                    await flushEvent.WaitAsync(token).ConfigureAwait(false);
+                else
+                {
+                    using (token.Register(() => diskRequest.asyncOperation.TrySetCanceled()))
+                        diskRequest = await diskRequest.asyncOperation.Task.WithCancellationAsync(token).ConfigureAwait(false);
+                }
+            }
+            catch (Exception e)
+            {
+                exceptionDispatchInfo = ExceptionDispatchInfo.Capture(e);
+            }
+
+            return new RmwAsyncResult<Input, Output, Context>(@this, fasterSession, currentCtx, pcontext, diskRequest, exceptionDispatchInfo);
+        }
+    }
+}
